+++ src/types/rx-document.d.ts
@@ -70,7 +70,9 @@ export declare interface RxDocumentBase<RxDocumentType, OrmMethods = {}> {
     destroy(): void;
 }
 
-export declare interface RxLocalDocument<Parent> extends RxDocumentBase<{}> {
+declare type LocalDocWithType<LocalDocType> = RxDocumentBase<LocalDocType> & LocalDocType;
+
+export declare type RxLocalDocument<Parent, LocalDocType = any> = RxDocumentBase<LocalDocType> & LocalDocType & {
     readonly parent: Parent;
     isLocal(): true;
 }